cyberpunk-eventos-backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
sqlalchemy>=2.0.36
python-multipart>=0.0.12
python-jose[cryptography]>=3.3.0
argon2-cffi>=25.1.0
python-dotenv>=1.0.1
pillow>=11.0.0
aiosqlite>=0.20.0
pydantic>=2.10.0
pydantic-settings>=2.6.0
